Move off of visual studio. Start transition to CMAKE. Move the 6502 emulator to it's on library so it can be repurposed.
In order to be more portable, the project will switch to CMAKE instead of the hell of visual studio. Side note: I hate development on windows.

Move the 6502 emulator so it can be developed separately from the APPLE II emulator and also so it can be used in multiple different emulator projects.